ov027: promote daPgDfdr_c into a single C++ TU
The penguin-defender level object becomes one production source,
src/actors/daPgDfdr_c.cpp, absorbing the fifteen one-function shards that
used to carry it. The ROM build takes all fifteen functions from a single
object, the way the cartridge's own build did.

WHAT THE CARTRIDGE PROVES ABOUT THE CLASS

  _ZTS  ov027 0x021139f0  "10daPgDfdr_c"
  _ZTI  ov027 0x021139e4  __si_class_type_info; its three words are
                          0x0209a764, 0x021139f0, 0x021089ec, and that
                          +8 word is _ZTI10dBgActor_c (ov002), so the
                          DIRECT base is dBgActor_c and nothing else
  _ZTV  ov027 0x02113a90  32 slots; slot 32 would be the 0x0209a764 that
                          opens the next RTTI record, so the table stops
                          at 31
  size  0x3dc             `mov r0, #0x3dc` in daPgDfdr_c_classInit
                          (ov027 0x0211207c) ahead of fBase_c::operator new

THE VTABLE was read out of the overlay slot by slot. Seven words point into
ov027: slots 0/3/6/9/12 (InitResources, CleanupResources, Behavior, Render,
OnPendingDestroy) and the 16/17 destructor pair. Slot 31 holds 0x020ee55c,
dBgActor_c's own Kill, so it is inherited and not overridden; every other
word is dActor_c's.

THE DESTRUCTOR MOVED INLINE INTO THE HEADER. The two shards each carried
their own out-of-line `daPgDfdr_c::~daPgDfdr_c() {}`; one TU can carry only
one, and an out-of-line one makes mwccarm emit D0 before D1 -- the reverse
of the cartridge's 0x021118c8 then 0x02111924 -- plus a homeless D2. Defined
in the class body it gives the retail pair in ROM order and no D2. That
moves the key function to InitResources, the first declared non-inline
virtual, so this TU still emits the _ZTV/_ZTI/_ZTS group. None of that data
lies inside the licensed .text range 0x021118c8..0x0211207c, so the ordinary
text-only route plus a twelve-row compiler_only_output block is correct;
dsd supplies the cartridge's bytes and production isolation discards the
emitted duplicates. The block licenses this class's own trio, the inherited
dBgActor_c/dActor_c/dBase_c/fBase_c chain, and Vector3's inline destructor
-- the one deadstrip-duplicate, because it is a function with a real ROM
home at arm9:0x020072c0. Every canonical address was read out of the
module's own symbols.txt.

ONE MERGE CHANGE COST A MATCH AND IS WORTH RECORDING. common.h and
math/Matrix.h both define Matrix4x3, both 0x30 bytes, and whichever a TU
sees first stands. The shard for func_ov027_02111994 included only common.h
and got the flat `s32 m[12]`; the merged file reached math/Matrix.h first
through daPgDfdr_c.h -> Model.h and got `Matrix3x3 r; Vector3 t`. Against
the nested spelling mwccarm splits that function's one whole-matrix
assignment into a 9-word ldm/stm run plus a 3-word tail it then CSEs against
the three `>> 3` stores above it: 0x94 -> 0xac, six instructions long.
Putting `#include "common.h"` first restores the cartridge's three flat
4-word ldm/stm pairs. That single line is the difference between 14/15 and
15/15.

NINE SYMBOLS THE SHARDS SPELLED TWO WAYS NOW HAVE ONE SPELLING, each
reconciled to the spelling whose address arithmetic is unchanged and every
call site adjusted with a cast rather than by changing what is computed.
include/decl_common.h's `(void*)` and `(void*,int)` settled the two helper
signatures it already declared. The five conflict notes tubuild froze into
the manifest are rewritten as RESOLVED and check_tubuild_conflicts agrees.

GATES

  tubuild verify    15/15 MATCH, objisolate clean, reloc-destinations
                    clean -> TEXT-VERIFIED, no unlicensed output
  tubuild linkcheck SCRATCH-LINK-VERIFIED (run before promotion); the
                    scratch ROM's sha256 is IDENTICAL to the baseline
                    control's, and dsd check symbols --fail has 0 new
                    errors against that control's 9 pre-existing ones
  rombuild -j 6     106/106 exact, 100.000000%; 11,088 source-built
                    functions reproducing, 0 mismatching
  romdata_check     11 data symbols from this source: 7 VERIFIED,
                    4 PARTIAL, 0 DIFFERS
  eligible          all 15 members plus daPgDfdr_c_classInit eligible
  layout_check, check_duplicate_sources, check_dead_references,
  check_tubuild_conflicts, check_header_offsets, port_refcheck,
  cpp_tu_compat --require-ready: all clean
  converted-baseline re-banked with tiers_ratchet --update (2597 -> 2597);
  the three removals are logged as the consolidation they are

Also adds notes/data/class-facts/daPgDfdr_c.json, the scout stage's ROM
evidence for this class.
