--- src/_ZN10daPgDfdr_c16OnPendingDestroyEv.cpp
@@ -1,10 +1,0 @@
-//cpp
-// @symbol _ZN10daPgDfdr_c16OnPendingDestroyEv
-/* daPgDfdr_c::OnPendingDestroy -- vtable slot 12. The ROM body is empty: the
- * override exists only to occupy the slot, so the base's own handler does not
- * run for this class. */
-#include "daPgDfdr_c.h"
-
-void daPgDfdr_c::OnPendingDestroy()
-{
-}
